made profile routes private and added current route logic

+++ controllers/profileControllers.js
@@ -1,16 +1,17 @@
 const asycHandler = require("express-async-handler");
 const Profile = require("../models/profileModel");
+
 //@desc Getting all profiles
 //@route GET /api/profiles
-//@access public
+//@access private
 const getAllProfiles = asycHandler(async (req, res) => {
-  const profiles = await Profile.find();
+  const profiles = await Profile.find({ user_id: req.user.id });
   res.status(200).json(profiles);
 });
 
 //@desc Creating a profile
 //@route POST /api/profiles
-//@access public
+//@access private
 const createProfile = asycHandler(async (req, res) => {
   console.log("request body is: ", req.body);
   const { name, email, phone } = req.body;
@@ -22,13 +23,14 @@ const createProfile = asycHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(200).json(profile);
 });
 
 //@desc Getting profile with specific id
 //@route GET /api/profiles/:id
-//@access public
+//@access private
 const getProfileByID = asycHandler(async (req, res) => {
   const profile = await Profile.findById(req.params.id);
   if (!profile) {
@@ -40,12 +42,16 @@ const getProfileByID = asycHandler(async (req, res) => {
 
 //@desc Update profile with id
 //@route PUT /api/profiles/:id
-//@access public
+//@access private
 const updateProfile = asycHandler(async (req, res) => {
   const profile = await Profile.findById(req.params.id);
   if (!profile) {
     res.status(404);
     throw new Error("Profile not found");
+  }
+  if (profile.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("Forbidden, trying to update contact of another user.");
   }
   const updatedProfile = await Profile.findByIdAndUpdate(
     req.params.id,
@@ -57,12 +63,16 @@ const updateProfile = asycHandler(async (req, res) => {
 
 //@desc Delete profile with id
 //@route DELETE /api/profile/:id
-//@access public
+//@access private
 const deleteProfile = asycHandler(async (req, res) => {
   const profile = await Profile.findById(req.params.id);
   if (!profile) {
     res.status(404);
     throw new Error("Profile not found");
+  }
+  if (profile.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("Forbidden, trying to delete contact of another user.");
   }
   await Profile.findByIdAndDelete(req.params.id);
   res.status(200).json(profile);
